task#89079 药品管理-同步优化 -药品新增平台业务代码编写

+++ ehealth-recipe-dao/src/main/java/recipe/dao/DrugListMatchDAO.java
@@ -211,7 +211,7 @@ public abstract class DrugListMatchDAO extends HibernateSupportDelegateDAO<DrugL
     public abstract List<DrugListMatch> findDataByOrganDrugCode(@DAOParam("organDrugCode")String organDrugCode);
 
     @DAOMethod(sql = "SELECT COUNT(*) from DrugListMatch where sourceOrgan =:sourceOrgan and status !=2 ")
-    public abstract Integer getNumBySourceOrgan(@DAOParam("sourceOrgan") Integer organDrugCode);
+    public abstract Long getNumBySourceOrgan(@DAOParam("sourceOrgan") Integer organDrugCode);
 
 
     /**
